[step1] feat: application 생성, 클래스 생성
