+++ README.md
@@ -41,7 +41,7 @@
   }
   ```
 - [x] 숫자 이외의 값을 전달할 경우 RuntimeException 예외가 발생해야 한다. (예 : “kimdabomi”)
-- [ ] 음수를 전달할 경우 RuntimeException 예외가 발생해야 한다. (예 : “-1,2,3”)
+- [x] 음수를 전달할 경우 RuntimeException 예외가 발생해야 한다. (예 : “-1,2,3”)
   ``` kotlin
   class StringAddCalculatorTest {
       private lateinit var calculator: StringAddCalculator
