+++ src/test/kotlin/calculator/StringAdditionCalculatorTest.kt
@@ -1,22 +1,8 @@
+package calculator
+
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.StringSpec
 import io.kotest.matchers.shouldBe
-
-class StringAdditionCalculator(private val str: String?) {
-    fun add(): Int {
-        if (str.isNullOrBlank()) return 0
-        val matchResult = Regex("//(.)\n(.*)").find(str)
-        val separator = matchResult?.groupValues?.get(1)
-        val numbers = matchResult?.groupValues?.get(2)?.split(",|:|${separator?.let { Regex.escape(it) }}".toRegex()) ?: str.split(",|:".toRegex())
-        if (numbers.any { it.matches(NOT_NUMBER_REGEX) }) throw RuntimeException(NOT_NUMBER_MESSAGE)
-        return if (numbers.size == 1 && separator == null) str.toInt() else numbers.sumOf { it.toInt() }
-    }
-
-    companion object {
-        private val NOT_NUMBER_REGEX = Regex("[^0-9]")
-        private const val NOT_NUMBER_MESSAGE = "숫자 이외의 값을 입력할 수 없습니다."
-    }
-}
 
 class StringAdditionCalculatorTest : StringSpec ({
     "빈 문자열 또는 null을 입력할 경우 0을 반환해야 한다." {
@@ -53,5 +39,12 @@ class StringAdditionCalculatorTest : StringSpec ({
             StringAdditionCalculator("1,2,a").add()
         }
         exception.message shouldBe "숫자 이외의 값을 입력할 수 없습니다."
+    }
+
+    "음수를 전달할 경우 RuntimeException 예외가 발생해야 한다." {
+        val exception = shouldThrow<RuntimeException> {
+            StringAdditionCalculator("-1,2").add()
+        }
+        exception.message shouldBe "음수를 입력할 수 없습니다."
     }
 })
